refactor: move LoopIndexVars.js -> generate/loop-index-vars.js
