+++ packages/compile/src/generate/equation-gen.js
@@ -4,7 +4,7 @@ import XLSX from 'xlsx'
 import { ModelLexer, ModelParser } from 'antlr4-vensim'
 import ModelReader from './ModelReader.js'
 import ModelLHSReader from './ModelLHSReader.js'
-import LoopIndexVars from './LoopIndexVars.js'
+import LoopIndexVars from './loop-index-vars.js'
 import Model from './Model.js'
 import {
   dimensionNames,
