+++ packages/compile/src/generate/loop-index-vars.js
